requirements.txt
# Python requirements for Vercel serverless functions
fastapi>=0.109.0
python-multipart>=0.0.7
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
httpx>=0.27.0
pydantic>=2.0
pydantic-settings>=2.0
pandas>=2.0
numpy>=1.26
yfinance>=0.2.30
python-dotenv>=1.0
tenacity>=8.0
structlog>=24.0
pypdf>=4.0.0
PyJWT>=2.8.0
supabase>=2.0.0
